feat: Add cloud function to add user to queue
This commit adds a new cloud function called `addUserToQueue` which allows users to be added to a queue. The function takes in parameters such as queue name, user name, public key, sats, and comment. It performs validation checks for required fields and then adds the user to the specified queue in the database. Additionally, it updates the total sats for the queue.